requirements.txt
nextcord
requests
python-dotenv
carter-py